requirements.txt
flask
numpy
pandas
scikit-learn
keras
tensorflow==2.11.0 \
  -f https://storage.googleapis.com/tensorflow/linux/cpu/ \
  tensorflow_cpu-2.11.0-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
tensorflow-io
azure-storage-blob
pika
boto3
flask_swagger_ui
flask_cors